// FRONTEND/chatbot-admin-panel/src/services/api/schemas/apiClientSettingsSchema.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { ApiClientSettingsSchemaHumanHandoffAgentGroupId } from './apiClientSettingsSchemaHumanHandoffAgentGroupId';
import type { ApiClientSettingsSchemaDefaultLlmModelConfigIdOverride } from './apiClientSettingsSchemaDefaultLlmModelConfigIdOverride';
import type { ApiClientSettingsSchemaDefaultVirtualAgentProfileIdOverride } from './apiClientSettingsSchemaDefaultVirtualAgentProfileIdOverride';
import type { ApiClientSettingsSchemaMaxTokensPerResponseOverride } from './apiClientSettingsSchemaMaxTokensPerResponseOverride';

export interface ApiClientSettingsSchema {
  /**
   * @minLength 3
   * @maxLength 100
   */
  application_id: string;
  allowed_context_ids?: number[];
  is_web_client?: boolean;
  allowed_web_origins?: string[];
  human_handoff_agent_group_id?: ApiClientSettingsSchemaHumanHandoffAgentGroupId;
  default_llm_model_config_id_override?: ApiClientSettingsSchemaDefaultLlmModelConfigIdOverride;
  default_virtual_agent_profile_id_override?: ApiClientSettingsSchemaDefaultVirtualAgentProfileIdOverride;
  /**
   * @minimum 0
   * @maximum 50
   */
  history_k_messages?: number;
  max_tokens_per_response_override?: ApiClientSettingsSchemaMaxTokensPerResponseOverride;
}