// FRONTEND/chatbot-admin-panel/src/services/api/schemas/lLMModelConfigResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { LLMProviderType } from './lLMProviderType';
import type { LLMModelType } from './lLMModelType';
import type { LLMModelConfigResponseBaseUrl } from './lLMModelConfigResponseBaseUrl';
import type { LLMModelConfigResponseDefaultTemperature } from './lLMModelConfigResponseDefaultTemperature';
import type { LLMModelConfigResponseDefaultMaxTokens } from './lLMModelConfigResponseDefaultMaxTokens';
import type { LLMModelConfigResponseConfigJson } from './lLMModelConfigResponseConfigJson';

export interface LLMModelConfigResponse {
  /**
   * @minLength 1
   * @maxLength 255
   */
  model_identifier: string;
  /**
   * @minLength 3
   * @maxLength 150
   */
  display_name: string;
  provider: LLMProviderType;
  model_type?: LLMModelType;
  is_active?: boolean;
  base_url?: LLMModelConfigResponseBaseUrl;
  default_temperature?: LLMModelConfigResponseDefaultTemperature;
  default_max_tokens?: LLMModelConfigResponseDefaultMaxTokens;
  supports_system_prompt?: boolean;
  config_json?: LLMModelConfigResponseConfigJson;
  id: number;
  created_at: string;
  updated_at: string;
  has_api_key?: boolean;
}