// FRONTEND/chatbot-admin-panel/src/services/api/schemas/humanAgentGroupResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { HumanAgentGroupResponseDescription } from './humanAgentGroupResponseDescription';

export interface HumanAgentGroupResponse {
  /**
   * @minLength 3
   * @maxLength 100
   */
  name: string;
  description?: HumanAgentGroupResponseDescription;
  is_active?: boolean;
  id: number;
}